src/protocols/adk_bridge.py
#!/usr/bin/env python3
"""
ADK Bridge - Bridge between TypeScript and Google's Agent Development Kit (ADK)

This script provides a bridge between TypeScript and Python for using Google's ADK.
It receives JSON-RPC style requests from stdin and sends responses to stdout.
"""

import json
import sys
import traceback
import asyncio
import logging
from typing import Dict, Any, Optional, List, Callable
from datetime import datetime

# Set up logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler("adk_bridge.log"),
        logging.StreamHandler(sys.stderr)
    ]
)
logger = logging.getLogger("adk_bridge")

# Import ADK if available
try:
    from google.adk import BaseAgent, Protocol
    from google.adk.tool import Tool, ToolSpec
    HAS_ADK = True
except ImportError:
    logger.warning("Google ADK not found. Running in compatibility mode.")
    HAS_ADK = False
    # Define stub classes for compatibility
    class BaseAgent:
        pass
    
    class Protocol:
        def __init__(self):
            pass
        
        async def register(self, agent):
            pass
        
        async def send_message(self, target, payload):
            pass
        
        def on_message(self, handler):
            pass
        
        async def update_context(self, context):
            pass
    
    class Tool:
        def __init__(self, name, description, function):
            self.name = name
            self.description = description
            self.function = function
    
    class ToolSpec:
        def __init__(self, name, description, parameters=None, returns=None):
            self.name = name
            self.description = description
            self.parameters = parameters or {}
            self.returns = returns or {}


class ADKBridge:
    """Bridge between TypeScript and Google's ADK"""
    
    def __init__(self):
        """Initialize the bridge"""
        self.initialized = False
        self.agent = None
        self.protocol = None
        self.tools = {}
        self.message_handlers = {}
        self.context = {}
    
    async def initialize(self) -> Dict[str, Any]:
        """Initialize the bridge"""
        if self.initialized:
            return {"status": "already_initialized"}
        
        try:
            # Create protocol instance
            self.protocol = Protocol()
            
            # Create agent instance
            self.agent = NewFuseAgent(self.protocol)
            
            # Register default message handler
            self.protocol.on_message(self.handle_message)
            
            self.initialized = True
            logger.info("ADK Bridge initialized successfully")
            return {"status": "initialized"}
        except Exception as e:
            logger.error(f"Failed to initialize ADK Bridge: {str(e)}")
            logger.error(traceback.format_exc())
            return {"status": "error", "error": str(e)}
    
    async def terminate(self) -> Dict[str, Any]:
        """Terminate the bridge"""
        self.initialized = False
        logger.info("ADK Bridge terminated")
        return {"status": "terminated"}
    
    async def register_tool(self, name: str, description: str, parameters: Dict[str, Any]) -> Dict[str, Any]:
        """Register a tool with the agent"""
        if not self.initialized:
            return {"status": "error", "error": "Bridge not initialized"}
        
        try:
            # Create tool spec
            tool_spec = ToolSpec(
                name=name,
                description=description,
                parameters=parameters
            )
            
            # Create tool function
            async def tool_function(**kwargs):
                # Send tool invocation to TypeScript
                result = await self.send_to_typescript("tool_invoked", {
                    "name": name,
                    "parameters": kwargs
                })
                return result
            
            # Create tool
            tool = Tool(
                name=name,
                description=description,
                function=tool_function
            )
            
            # Register tool with agent
            self.agent.register_tool(tool)
            
            # Store tool for later reference
            self.tools[name] = tool
            
            logger.info(f"Registered tool: {name}")
            return {"status": "success", "tool": name}
        except Exception as e:
            logger.error(f"Failed to register tool {name}: {str(e)}")
            logger.error(traceback.format_exc())
            return {"status": "error", "error": str(e)}
    
    async def register_message_handler(self, message_type: str, handler_id: str) -> Dict[str, Any]:
        """Register a message handler"""
        if not self.initialized:
            return {"status": "error", "error": "Bridge not initialized"}
        
        try:
            # Store handler ID for later reference
            self.message_handlers[message_type] = handler_id
            
            logger.info(f"Registered message handler for {message_type}: {handler_id}")
            return {"status": "success", "message_type": message_type, "handler_id": handler_id}
        except Exception as e:
            logger.error(f"Failed to register message handler for {message_type}: {str(e)}")
            logger.error(traceback.format_exc())
            return {"status": "error", "error": str(e)}
    
    async def send_message(self, target: str, message_type: str, payload: Dict[str, Any]) -> Dict[str, Any]:
        """Send a message to another agent"""
        if not self.initialized:
            return {"status": "error", "error": "Bridge not initialized"}
        
        try:
            # Create message
            message = {
                "type": message_type,
                "payload": payload,
                "timestamp": datetime.now().timestamp(),
                "sender": self.agent.id
            }
            
            # Send message
            await self.protocol.send_message(target, message)
            
            logger.info(f"Sent message to {target}: {message_type}")
            return {"status": "success", "target": target, "message_type": message_type}
        except Exception as e:
            logger.error(f"Failed to send message to {target}: {str(e)}")
            logger.error(traceback.format_exc())
            return {"status": "error", "error": str(e)}
    
    async def update_context(self, context: Dict[str, Any]) -> Dict[str, Any]:
        """Update the agent context"""
        if not self.initialized:
            return {"status": "error", "error": "Bridge not initialized"}
        
        try:
            # Update context
            self.context.update(context)
            
            # Update protocol context
            await self.protocol.update_context(self.context)
            
            logger.info(f"Updated context: {context.keys()}")
            return {"status": "success", "context_keys": list(context.keys())}
        except Exception as e:
            logger.error(f"Failed to update context: {str(e)}")
            logger.error(traceback.format_exc())
            return {"status": "error", "error": str(e)}
    
    async def handle_message(self, message: Dict[str, Any]) -> None:
        """Handle a message from another agent"""
        try:
            message_type = message.get("type")
            
            # Check if we have a handler for this message type
            handler_id = self.message_handlers.get(message_type)
            if handler_id:
                # Send message to TypeScript
                await self.send_to_typescript("message_received", {
                    "handler_id": handler_id,
                    "message": message
                })
            else:
                logger.warning(f"No handler for message type: {message_type}")
        except Exception as e:
            logger.error(f"Failed to handle message: {str(e)}")
            logger.error(traceback.format_exc())
    
    async def send_to_typescript(self, event: str, data: Dict[str, Any]) -> Any:
        """Send an event to TypeScript and wait for a response"""
        # This would be implemented to send data back to TypeScript
        # For now, we'll just log it
        logger.info(f"Would send to TypeScript: {event} - {data}")
        return {"status": "success"}


class NewFuseAgent(BaseAgent):
    """Agent implementation for The New Fuse"""
    
    def __init__(self, protocol: Protocol):
        """Initialize the agent"""
        super().__init__()
        self.id = "new_fuse_agent"
        self.protocol = protocol
        self.tools = {}
    
    def register_tool(self, tool: Tool) -> None:
        """Register a tool with the agent"""
        self.tools[tool.name] = tool


# Main bridge instance
bridge = ADKBridge()

async def process_request(request: Dict[str, Any]) -> Dict[str, Any]:
    """Process a request from TypeScript"""
    request_id = request.get("id")
    method = request.get("method")
    args = request.get("args", {})
    
    try:
        # Check if method exists
        if not hasattr(bridge, method) or not callable(getattr(bridge, method)):
            return {
                "id": request_id,
                "error": f"Unknown method: {method}"
            }
        
        # Call method
        method_func = getattr(bridge, method)
        result = await method_func(**args)
        
        return {
            "id": request_id,
            "result": result
        }
    except Exception as e:
        logger.error(f"Error processing request: {str(e)}")
        logger.error(traceback.format_exc())
        return {
            "id": request_id,
            "error": str(e)
        }


async def main() -> None:
    """Main entry point"""
    logger.info("ADK Bridge starting")
    
    # Process stdin for requests
    for line in sys.stdin:
        try:
            # Parse request
            request = json.loads(line.strip())
            
            # Process request
            response = await process_request(request)
            
            # Send response
            sys.stdout.write(json.dumps(response) + "\n")
            sys.stdout.flush()
        except json.JSONDecodeError:
            logger.error(f"Failed to parse request: {line.strip()}")
        except Exception as e:
            logger.error(f"Unexpected error: {str(e)}")
            logger.error(traceback.format_exc())


if __name__ == "__main__":
    asyncio.run(main())
