requirements.txt
google-adk
google-a2a
